% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregacja_adm.R
\name{agreguj_1rokpo_adm}
\alias{agreguj_1rokpo_adm}
\title{Obliczanie wskaźników na poziomie zagregowanym dla monitoringu karier
na danych administracyjnych}
\usage{
agreguj_1rokpo_adm(wsk2, wsk3, wsk4, podzial_grupy, rok_abso)
}
\arguments{
\item{wsk2}{ramka danych z tabeli pośredniej nr 2 (P2) z wynikami z 4.
rundy monitoringu na danych administracyjnych}

\item{wsk3}{ramka danych z tabeli pośredniej nr 3 (P3) z wynikami z 4.
rundy monitoringu na danych administracyjnych}

\item{wsk4}{ramka danych z tabeli pośredniej nr 4 (P4) z wynikami z 4.
rundy monitoringu na danych administracyjnych}

\item{podzial_grupy}{ramka danych zawierająca definicje podziałów na grupy -
np. zwrócona przez funkcję \code{\link{utworz_grupowanie_ze_zmiennej}}}

\item{rok_abso}{rok, w którym grupa absolwentów uzyskała status absolwenta.
Może to być więcej niż 1 wartość.}
}
\value{
list
}
\description{
Funkcja obliczająca wskaźniki na poziomie zagregowanym na
potrzeby raportów \verb{1rokpo} (po roku od ukończenia szkoły).
}
\seealso{
\code{\link{agreguj_wskazniki}} oraz przekazywane do niej funkcje
używane do obliczania konkretnych wskaźników zagregowanych:
\itemize{
\item{\code{\link{dane_szkoly}},}
\item{\code{\link{l_abs}},}
\item{\code{\link{l_kobiet}},}
\item{\code{\link{l_abs_zrodla}},}
\item{\code{\link{status_S3_mies}},}
\item{\code{\link{zawody_status_S3}},}
\item{\code{\link{E2_nauka_kontyn}},}
\item{\code{\link{Z4_ods_prac_mies}},}
\item{\code{\link{Z8_formy_prac_mies}},}
\item{\code{\link{Z9_kont_mlod}},}
\item{\code{\link{W3_sr_doch_uop}},}
\item{\code{\link{B2_ods_bezrob}},}
\item{\code{\link{liczebnosc_branze_ucz}},}
\item{\code{\link{rozklad_liczebnosc}},}
\item{\code{\link{rozklad_zawody}}}
}
}
