% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{B2_ods_bezrob}
\alias{B2_ods_bezrob}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
B2_ods_bezrob(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
list
}
\description{
Funkcja licząca na potrzeby szablonu raportu rozkład liczby
miesięcy bezrobocia rejestrowanego wśród absolwentów od września do grudnia w
roku ukończenia szkoły.
}
