% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprawdzanie_danych.R
\name{check_kolumny_tabele}
\alias{check_kolumny_tabele}
\title{Sprawdzanie poprawności danych}
\usage{
check_kolumny_tabele(
  data,
  tabela = c("p1", "p2", "p3", "p4", "p5"),
  force = FALSE
)
}
\arguments{
\item{data}{ramka danych, której poprawność nazw kolumn ma być sprawdzona.
Zwykle jest to tabela pośrednia w rozumieniu tabel używanych w Monitoringu
Karier}

\item{tabela}{wartość tekstowa oznaczająca która tabela pośrednia ma być
sprawdzona}

\item{force}{wartość logiczna: \code{FALSE} - funkcja zatrzymuje swoje działanie
po napotkaniu błędu, \code{TRUE} - funkcja kontynuuje działanie mimo napotkania
błędu}
}
\description{
Funkcja sprawdza czy dana tabela pośrednia posiada kolumny, które są kluczowe
dla poprawnego przetwarzania tabel pośrednich oraz do generowania wskaźników
zagregowanych.
}
