% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_zagregowane.R
\name{definicje_podzialu}
\alias{definicje_podzialu}
\title{Funkcje liczące wskaźniki zagregowane - Monitoring Karier}
\usage{
definicje_podzialu(p4, rodzaj_wsk, rok_ukonczenia = 2024)
}
\arguments{
\item{p4}{tabela pośrednia \emph{p4}}

\item{rodzaj_wsk}{ciąg znaków określający dla jakiego rodzaju wskaźników
należy zwórócić definicję podziału na grupy. Możliwe wartości:
\describe{
\item{\code{"szk_god"}}{Główny podział na szkołę oraz jej grupę odniesienia z
wykluczeniem}
\item{\code{"god1_god2"}}{}
\item{\code{"szkozaw"}}{Zbiór, gdzie jednostką podziału jest zawód nauczany (i
posiadający absolwentów) w danej szkol - szkoło-zawod.}
\item{\code{"woj"}}{}
}}

\item{rok_ukonczenia}{rok, którym absolwent ukończył szkołę (jest to tym
samym rok monitoringu)}
}
\value{
ramka danych będąca zbiorem definicji podziałów na agregaty
}
\description{
Funkcja służąca do wygenerowania ramki danych będącej zbiorem definicji
podziałów na grupy i ewentualnie grupy odniesienia dla danego zestawu
wskaźników zagregowanych.
}
\seealso{
\code{\link[=utworz_grupowanie_odn_teryt]{utworz_grupowanie_odn_teryt()}}
}
