% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_zagregowane.R
\name{dodaj_odmiany}
\alias{dodaj_odmiany}
\title{Dodawanie do zbioru zmiennej zawierającej odmiany np. rzeczowników}
\usage{
dodaj_odmiany(wskazniki, rodzaj_wsk)
}
\arguments{
\item{wskazniki}{zbiór wskaźników zagregowanch (\code{szk} lub \code{woj})}

\item{rodzaj_wsk}{określa dla jakiego zbioru wyliczyć odmiany}
}
\value{
zbiór wskaźników zagregowanch z dołączoną zmienną \code{odmiany}
}
\description{
Funkcja używana w ramach funkcji \code{\link[=zlacz_partial]{zlacz_partial()}}. Na podstawie zmiennych ze
zbioru wskaźników zagregowanch \code{szk} lub \code{woj} określa odmianę rzeczowników,
które następnie używane są w raportach szkolnych.
}
