% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_zagregowane.R
\name{dodaj_odmiany_szk}
\alias{dodaj_odmiany_szk}
\title{Dodawanie do zbioru \code{szk} zmiennej zawierającej odmiany rzeczowników}
\usage{
dodaj_odmiany_szk(obiekt_szk)
}
\arguments{
\item{obiekt_szk}{zbiór wskaźników zagregowanch \code{szk}}
}
\value{
zbiór wskaźników zagregowanch \code{szk} z dołączoną zmienną \code{odmiany}
}
\description{
Funkcja używana w ramach funkcji \code{\link[=zlacz_partial]{zlacz_partial()}}. Na podstawie zmiennych ze
zbioru wskaźników zagregowanch \code{szk} określa odmianę rzeczowników, które
następnie używane są w raportach szkolnych.
}
