% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{E2_nauka_kontyn}
\alias{E2_nauka_kontyn}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
E2_nauka_kontyn(x, rok, mies = 12)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{rok}{rok osiągnięcia statusu absolwenta}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik - domyślnie
grudzień}
}
\value{
list
}
\description{
Funkcja licząca wskaźnik E2 - Sposoby kontynuowania edukacji.
Domyślnym okresem, w którym sprawdzany jest odsetek kontynuujących naukę jest
grudzień roku ukończenia szkoły (\code{rok_abs}). Funkcja liczy wskaźnik dla
pojedynczego okresu, a nie dla zakresu dat.
}
