% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_wskazniki_adm.R
\name{ind_Z8_formy_prac}
\alias{ind_Z8_formy_prac}
\title{Wskaźniki indywidualne na potrzeby monitoringu karier - dane
administracyjne}
\usage{
ind_Z8_formy_prac(x)
}
\arguments{
\item{x}{zbiór osobo-miesięcy (zwykle ramka danych pośrednich P3)}
}
\value{
vector
}
\description{
Funkcja obliczająca formę wykonywanej pracy na podstawie
zmiennej \code{praca} z tabeli pośredniej \code{P3}.
Wyróżnione formy pracy:
\itemize{
\item{\code{ucz_uop}}{Zatrudnienie na podstawie UOP oraz brak innej formy},
\item{\code{ucz_samoz}}{Prowadzący działalność gosp. (samozatrudnienie)
oraz brak innej formy},
\item{\code{ucz_inna}}{Tylko forma inna niż samozatrudnienie i niż umowa o
pracę (np. umowa-zlecenie)},
\item{\code{ucz_wiecej}}{Więcej niż jedna forma, czyli: Zatrudnienie na
podstawie UOP oraz prowadzący działalność gosp., Zatrudnienie na podstawie
UOP oraz zatrudnienie w innej formie, Prowadzący działalność gosp.
(samozatrudnienie) oraz zatrudnienie w innej formie, Zatrudnienie na
podstawie UOP oraz działalność gosp. oraz zatrudnienie w innej formie}
\item{\code{NA}{Brak danych - dla absolwentów, którzy w zmiennej `praca`
 posiadają wartość `0` lub brak danych (czyli pozostałe wartości spoza
 zakresu od 1 do 7)}}
}
}
