% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{l_abs}
\alias{l_abs}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
l_abs(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
numeric
}
\description{
Funkcja oblicza wskaźnik opisujący liczbę absolwentów na danym
poziomie agregacji.
}
