% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{l_abs_zrodla}
\alias{l_abs_zrodla}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
l_abs_zrodla(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
list
}
\description{
Funkcja oblicza wskaźnik opisujący liczbę absolwentów na danym
poziomie agregacji, o których pozyskano informacje z poszczególnych
rejestrów.
}
