% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{l_kobiet}
\alias{l_kobiet}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
l_kobiet(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
numeric
}
\description{
Funkcja oblicza wskaźnik opisujący liczbę kobiet wśród
absolwentów, którzy zostali objęci monitoringiem na danym poziomie agregacji.
}
