% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{licz_zawody}
\alias{licz_zawody}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
licz_zawody(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
list
}
\description{
Funkcja licząca rozkład liczebności w zawodach na potrzeby
raportu wojewódzko-branżowego.
}
