% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{liczebnosc_branze_ucz}
\alias{liczebnosc_branze_ucz}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
liczebnosc_branze_ucz(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
list
}
\description{
Funkcja licząca rozkład liczebności wyuczonych zawodów wśród
absolwentów w podziale na branże. Dodatkowo, funkcja liczy liczebności
absolwentów w branżach, a informacja ta służy jako podstawa do definiowania
warunków w szablonie raportu dla szkół branżowych 1. stopnia.
}
\details{
UWAGA: Wynikiem działania funkcji jest rozkład liczebności, w którym
nie odfiltrowuje się obserwacji poniżej 10 absolwentów, więc należy to zrobić
na poziomie szablonu raportu.
}
