% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{liczebnosc_dziedziny}
\alias{liczebnosc_dziedziny}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
liczebnosc_dziedziny(x, ROK, mies = 12)
}
\arguments{
\item{x}{ramka danych zawierająca informację o kontynuowaniu
kształcenia w danej dziedzinie (tabela danych pośrednich P2 lub zawierająca
analogiczne informacje oraz te same nazwy kolumn co tabela P2)}

\item{ROK}{rok osiągnięcia statusu absolwenta}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik - domyślnie jest
to grudzień}
}
\value{
list
}
\description{
Funkcja licząca rozkład liczebności absolwentów kontynuujących
naukę na studiach w podziale na dziedziny. Funkcja liczy wskaźnik tylko dla
absolwentów techników,liceów ogólnokształcących i branżowych szkół 2.
stopnia.
}
