% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robocze_funkcje.R
\name{normal_windows_path}
\alias{normal_windows_path}
\title{Zmiana ukośników w ścieżce na Windowsowe}
\usage{
normal_windows_path(path)
}
\arguments{
\item{path}{Ścieżka pliku jako ciąg znaków (string). Musi być niepusty.}
}
\value{
Zwraca tekst z poprawionymi ukośnikami, gotowy do użycia w Windows.
}
\description{
Funkcja przyjmuje ścieżkę pliku jako ciąg znaków i zamienia wszystkie ukośniki
z formatu Unixowego (\code{/}) na format Windowsowy (\verb{\\\\}).#'
}
\examples{
normal_windows_path("folder/podfolder/plik.txt")
# [1] "folder\\podfolder\\plik.txt"
}
