% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robocze_funkcje.R
\name{okres_na_date}
\alias{okres_na_date}
\title{Przeliczanie zmiennej \code{okres} na miesiąc i rok}
\usage{
okres_na_date(x)
}
\arguments{
\item{x}{wartość liczbowa, która ma być przeliczona na miesiąc i rok
(zazwyczaj jest to zmienna \code{okres} w tabeli pośredniej)}
}
\value{
list
}
\description{
Funkcja przeliczająca liczbę miesięcy, które minęły od roku 0 na
miesiąc i rok.
}
