% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robocze_funkcje.R
\name{okres_na_rok}
\alias{okres_na_rok}
\title{Przeliczanie zmiennej \code{okres} na rok}
\usage{
okres_na_rok(x)
}
\arguments{
\item{x}{data wyrażona jako liczbą miesięcy, które minęły od roku 0, czyli od
początku naszej ery (zazwyczaj jest to zmienna \code{okres} w tabeli
pośredniej), która ma być przeliczona na wartość liczbową będącą rokiem}
}
\value{
numeric
}
\description{
Funkcja przeliczająca liczbę miesięcy, które minęły od roku 0 na
rok.
}
