% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_zagregowane.R
\name{plik_source}
\alias{plik_source}
\title{Funkcja generująca plik ze ścieżkami do skryptów}
\usage{
plik_source(sciezka_docelowa, rodzaj_wsk)
}
\arguments{
\item{sciezka_docelowa}{ścieżka w formacie tekstowym, w ktorej ma być
zapisany plik source - jest to równocześnie ścieżka, w której powinny
znajdować się skrypty według podziału, który ma być zastosowany w
zrównoleglaniu}

\item{rodzaj_wsk}{ciąg znaków określający dla jakiego rodzaju wskaźników
należy zwórócić definicję podziału na grupy. Możliwe wartości to \code{"szk_god"}
lub \code{"szkozaw"}. Dla poozstałych wartości funkcja nic nie zwraca.}
}
\value{
plik txt zawierający ścieżki do skryptów R liczących wskaźniki
zagregowane
}
\description{
Funkcja zwraca plik txt będący spisem ścieżek do poszczególnych skryptów,
które tworzone są za pomocą funkcji \code{\link[=skrypt_wzor]{skrypt_wzor()}} - mogą również być
stworzone w inny sposób, ale ważne żeby ich nazewnictwo i struktura były
zgodne z tymi, które zwraca wspomniana funkcja.
}
\seealso{
\code{\link[=skrypt_bat]{skrypt_bat()}}

\code{\link[=skrypt_wzor]{skrypt_wzor()}}
}
