% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabele_posrednie.R
\name{przygotuj_tabele_posrednie}
\alias{przygotuj_tabele_posrednie}
\title{Funkcje przekształcające surowe tabele pośrednie - Monitoring Karier - Edycja 5}
\usage{
przygotuj_tabele_posrednie(tabele_sciezka, rok_ukonczenia = 2024)
}
\arguments{
\item{tabele_sciezka}{ścieżka do tabel pośrednich}

\item{rok_ukonczenia}{rok, którym absolwent ukończył szkołę (jest to tym
samym rok monitoringu)}
}
\value{
multiple \code{.RData} objects
}
\description{
Funkcja przekształcająca surowe tabele pośrednie do formy gotowej do użytku w
dalszej części procesu przechodzenia od tabel pośrednich do generowania
automatycznych raportów.
}
\seealso{
\code{\link[=rozdziel_tabele]{rozdziel_tabele()}}
}
