% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{rozklad_liczebnosc}
\alias{rozklad_liczebnosc}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
rozklad_liczebnosc(x, ROK = 2024, mies = 12, zmienna, plc = NULL)
}
\arguments{
\item{x}{ramka danych zawierająca informację o kontynuowaniu
kształcenia w danej dziedzinie/dyscyplinie/branży (tabela danych pośrednich
P2 lub zawierająca analogiczne informacje oraz te same nazwy kolumn co tabela
P2)}

\item{ROK}{rok osiągnięcia statusu absolwenta}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik - domyślnie jest
to grudzień}

\item{zmienna}{zmienna, której rozkład ma zwrócic funkcja}

\item{plc}{płeć absolwenta przekazana jako wartość tekstowa ("K" lub "M") -
dla rozkładów przez płeć}
}
\value{
list
}
\description{
Funkcja licząca rozklad zmeinnej z tabeli P2 przekazanej w
argumencie funkcji. Zmienne, dla których zaprojektowano funkjcę to:
\itemize{
\item{\code{dyscyplina_wiodaca_kont}}{Dyscyplina studiów, w której absolwent
kontynuował naukę}
\item{\code{dziedzina_kont}}{Dziedzina studiów, w której absolwent
kontynuował naukę na studiach}
\item{\code{branza_kont}}{Branża zawodu KZSB, w której absolwent kontynuował
naukę w szkole kształcącej w zawodzie}
}
}
