% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{rozklad_zawody}
\alias{rozklad_zawody}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
rozklad_zawody(x, ROK = 2024, mies = 12, zmienna)
}
\arguments{
\item{x}{ramka danych zawierająca informację o kontynuowaniu
kształcenia w danej dyscyplinie lub branży (tabela danych pośrednich P2 lub
zawierająca analogiczne informacje oraz te same nazwy kolumn co tabela P2)}

\item{ROK}{rok osiągnięcia statusu absolwenta}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik - domyślnie jest
to grudzień}

\item{zmienna}{}
}
\value{
list
}
\description{
Funkcja licząca rozkład liczebności danej zmiennej (funkcję
zaprojektowano do pracy ze zmiennymi \code{p2$branza} oraz
\code{p2$dyscyplina_wiodaca_kont}) absolwentów w podziale na wykonywany zawód.
Funkcja liczy wskaźnik tylko dla absolwentów techników, szkół policealnych i
branżowych szkół 2. stopnia.
}
