% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_zagregowane.R
\name{skrypt_bat}
\alias{skrypt_bat}
\title{Tworzenie skryptów \code{.bat} uruchamiających zrównoleglanie obliczeń}
\usage{
skrypt_bat(sciezka_docelowa, rodzaj_wsk)
}
\arguments{
\item{sciezka_docelowa}{ścieżka w formacie tekstowym, w ktorej ma być
zapisany plik source - jest to równocześnie ścieżka, w której powinny
znajdować się skrypty według podziału, który ma być zastosowany w
zrównoleglaniu}

\item{rodzaj_wsk}{ciąg znaków określający dla jakiego rodzaju wskaźników
należy zwórócić definicję podziału na grupy. Możliwe wartości to \code{"szk_god"}
lub \code{"szkozaw"}. Dla poozstałych wartości funkcja nic nie zwraca.}
}
\value{
skrypt \code{.bat} uruchamiający skrypty w oddzielnych sesjach R
}
\description{
Funkcja zwraca skrypt \code{.bat}, który uruchamia tyle niezależnych sesji R, ile
wierszy zawiera plik "source" zawierający ścieżki do skryptów R służących
zrównolegleniu obliczeń.
}
