% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_zagregowane.R
\name{skrypty_podzial}
\alias{skrypty_podzial}
\title{Funkcja tworząca podział na skrypty}
\usage{
skrypty_podzial(nrow, podzial = 11)
}
\arguments{
\item{nrow}{liczba wierszy w ramce danych wygenerowanej za pomocą funkcji
\code{\link[=utworz_grupowanie_odn_teryt]{utworz_grupowanie_odn_teryt()}}}

\item{podzial}{liczba opisująca na ile części ma zostać podzielona liczba
wierszy (tyle będzie finalnie działających sesji R podczas zrównoleglania)}
}
\value{
ramka danych zawierająca zakres wierszy dla każdej jednostki podziału
}
\description{
Funkcja tworząca podział na skrypty
}
\details{
Funkcja zwraca ramkę danych będącą definicją podziału na zbliżoną
liczbę wierszy w amce danych z definicją podziału na grupy. Podział ten jest
potrzebny do zrównoleglenia obliczeń - do stworzenia poszczególnych skryptów
będących częścią infrastruktury zrównoleglającej obliczenia.
}
