% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{status_S3_mies}
\alias{status_S3_mies}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
status_S3_mies(x, rok_od, mies_od, rok_do, mies_do)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{rok_od}{rok początku okresu, dla którego ma być policzony wskaźnik}

\item{mies_od}{miesiąc początku okresu, dla którego ma być policzony wskaźnik}

\item{rok_do}{rok końca okresu, dla którego ma być policzony wskaźnik}

\item{mies_do}{miesiąc końca okresu, dla którego ma być policzony wskaźnik}
}
\value{
list
}
\description{
Funkcja licząca odsetek absolwentów o danym statusie
edukacyjno-zawodowym (wskaźnik S3) w danym miesiącu \code{mies} dla raportu dla
danego roku.
}
\details{
W tej edycji pojawia się nowy status - bezrobocie rejestrowane,
który został wyodrębniony ze statusu \code{neet}. Wyróżniamy następujące statusy
(bez KUZ i KKZ):
\itemize{
\item{ucz_prac}{nauka i praca}
\item{tylko_ucz}{tylko nauka}
\item{tylko_prac}{tylko praca}
\item{bezrob}{zarejestrowany jako bezrobotny}
\item{neet}{brak nauki i brak pracy lub brak informacji o tychże}
}
Bezrobotni kontynuujący naukę niestacjonarnie nie są wykluczani z kategorii
„Nauka”, więc określając ich status w danym miesiącu traktujemy ich jako
osoby uczące się. W sytuacji gdy w tym samym miesiącu absolwent miał różne
statusy, przypisujemy mu „bardziej aktywny” z tych statusów (np. „naukę” a
nie „bezrobocie”).

Zmienna \code{status} powinna mieć wartości w formacie skróconym, czyli powinny
one być stworzone za pomocą argumentu \code{etykiety = FALSE} funkcji
\link{\code{ind_status_S3}}, ponieważ nazwy statusów w tej formie są potem
używane w szablonie raportu szkolnego.
}
