% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabele_posrednie.R
\name{uzupelnij_wynagrodzenie_powiat}
\alias{uzupelnij_wynagrodzenie_powiat}
\title{Funkcje przekształcające surowe tabele pośrednie - Monitoring Karier - Edycja 5}
\usage{
uzupelnij_wynagrodzenie_powiat(
  tabela_p3,
  rok_ukonczenia = 2024,
  wynagrodzenia = NULL,
  force = FALSE
)
}
\arguments{
\item{tabela_p3}{ścieżka do pliku z tabelą pośrednią \code{p3}, która przechowuje
m.in. informację o średnim wynagrodzeniu w powiecie na podstawie danych z
GUS.}

\item{rok_ukonczenia}{rok, którym absolwent ukończył szkołę (jest to tym
samym rok monitoringu)}

\item{wynagrodzenia}{tabela z mapowaniem wynagrodzeń. Jest to argument
opcjonalny, ponieważ funkcja podejmie próbę uzupełnienia brakujących wartości
pobierając je z BDL, ale jeśli się to nie uda można przekazać jako argument
swoją tabelę z mapowaniem wynagrodzeń na teryty powiatów i na jej podstawie
brakujące wartości zostaną dołączone do tabeli pośredniej.}

\item{force}{wartość logiczna mówiąca o tym czy uzupełnianie wartości ma się
odbywać nawet jeśli nie ma braków danych}
}
\value{
tabela pośrednia \code{p3}
}
\description{
Funkcja uzupełnia brakujące dane w zmiennej \code{powiat_sr_wynagrodzenie} w
tabeli pośredniej \code{p3}. Do tej pory spotkaliśmy się z dwoma rodzajami braków
danych:
\itemize{
\item{brak \strong{wszystkich} wynagrodzeń}{czasem zdarza się, że z jakiegoś
powodu tabela \code{p3} w ogóle nie ma danych o wynagrodzeniach w powiecie. W
takim wypadku funkcja pobierze dane z BDL i dołączy je do zbioru.}
\item{brak \strong{wybranych} wynagrodzeń}{w ostatnich latach GUS przestał
publikować dane dla powiatu Jastrzębie-Zdrój (kod teryt: 2467). Do
poprawnego działania funkcji używanych w dalszych krokach procesu
generowania raportów automatycznych wymagana jest kompletność danych o
przeciętnym wynagrodzeniu w powiecie.}
}
Funkcja rozpoznaje rodzaj braku danych i uzupełnia je albo pobierając dane z
BDL albo przewidując brakującą wartość w danym roku na podstawie wartości z
lat poprzednich w powiecie oraz w województwie.
}
\details{
Funkcja, w zależności od napotkanych braków danych, podejmuje rózne próby
uzupełnienia danych. W pierwszym kroku sprawdzane jest czy w ogóle zmienna
dotycząca wynagrodzeń w powiatach zawiera jakieś braki danych. Jeżeli ich nie
ma, to działanie jest przerywane, natomiast można je wymusić za pomocą
argumentu \code{force = TRUE}.
}
