% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{W3_sr_doch_uop}
\alias{W3_sr_doch_uop}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
W3_sr_doch_uop(x, nauka)
}
\arguments{
\item{x}{ramka danych pośrednich P4}

\item{nauka}{wartość TRUE/FALSE określająca czy status ma być liczony dla
absolwentów uczących się czy nie uczących się}

\item{wynagrodzenie_powiaty}{ramka danych pośrednich P3}
}
\value{
list
}
\description{
Funkcja licząca średni względny miesięczny przychód z pracy
\strong{etatowej} (na Umowie o Pracę - UOP) w danym okresie w odniesieniu do
zarobków w zamieszkiwanym powiecie. Domyślnie jest to okres między wrześniem
a grudniem roku przekazanego w argumentach funkcji, ale te wartości można
zmieniać. Średnia liczona jest oddzielnie dla uczących i nie uczących się
absolwentów.
}
\details{
Relatywna średnia zarobków liczona jest w kilku etapach. W pierwszym
kroku, dla każdego absolwenta liczony jest stosunek wartości wynagrodzenia na
UOP w danym miesiącu do średniego wynagrodzenia w powiecie, w którym
uczęszczał do szkoły według GUS. Następnie, dla każdego absolwenta liczona
jest ze stosunków zarobków absolwenta w poszczególnych miesiącach do średnich
zarobków w powiecie. Ze zbioru osobo-miesięcy przechodzimy na zbiór osób,
gdzie każdy wiersz odpowiada absolwentowi, a iloraz zarobków absolwenta i
zarobków w powiecie jest uśredniony - średni odsetek dla wszystkich miesięcy,
w których dochód był niezerowy. Następnie na tych indywidualnych wartościach
liczone są statystyki opisowe dla zadanego poziomu agregacji.
}
