% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{Z4_ods_prac_mies}
\alias{Z4_ods_prac_mies}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
Z4_ods_prac_mies(x, nauka)
}
\arguments{
\item{x}{ramka danych pośrednich P4}

\item{nauka}{wartość TRUE/FALSE określająca czy status ma być liczony dla
absolwentów uczących się (TRUE) czy nie uczących się (FALSE)}
}
\value{
list
}
\description{
Funkcja licząca odsetek pracujących w okresie od września do
grudnia roku ukończenia szkoły w podziale na pobierających i nie
pobierających naukę.
}
