% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{Z8_formy_prac_mies}
\alias{Z8_formy_prac_mies}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
Z8_formy_prac_mies(x, rok, mies = 12, nauka)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{rok}{rok lub zakres lat osiągnięcia statusu absolwenta}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik}

\item{nauka}{wartość TRUE/FALSE określająca czy status ma być liczony dla
absolwentów uczących się czy nie uczących się}
}
\value{
list
}
\description{
Funkcja licząca odsetek absolwentów wykonujących dane formy
pracy w danym miesiącu (domyślnie jest to grudzień). Wskaźnik może być
liczony albo dla absolwentów pracujących i kontynuujących naukę (\code{nauka
= TRUE}) lub dla absolwentów pracujących i nie kontynuujących nauki
(\code{nauka = FALSE}). Wyróżnione formy pracy:
\itemize{
\item{\code{ucz_uop}}{Zatrudnienie na podstawie UOP oraz brak innej formy},
\item{\code{ucz_samoz}}{Prowadzący działalność gosp. (samozatrudnienie)
oraz brak innej formy},
\item{\code{ucz_inna}}{Tylko forma inna niż samozatrudnienie i niż umowa o
pracę (np. umowa-zlecenie)},
\item{\code{ucz_wiecej}}{Więcej niż jedna forma, czyli: Zatrudnienie na
podstawie UOP oraz prowadzący działalność gosp., Zatrudnienie na podstawie
UOP oraz zatrudnienie w innej formie, Prowadzący działalność gosp.
(samozatrudnienie) oraz zatrudnienie w innej formie, Zatrudnienie na
podstawie UOP oraz działalność gosp. oraz zatrudnienie w innej formie}
\item{\code{NA}{Brak danych - dla absolwentów, którzy w zmiennej `praca`
 posiadają wartość `0` lub brak danych (czyli pozostałe wartości spoza
 zakresu od 1 do 7)}}
}
}
