% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{Z9_kont_mlod}
\alias{Z9_kont_mlod}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
Z9_kont_mlod(x, rok, mies = 9, nauka)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{rok}{rok lub zakres lat osiągnięcia statusu absolwenta}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik}

\item{nauka}{wartość TRUE/FALSE określająca czy status ma być liczony dla
absolwentów uczących się (TRUE) czy nie uczących się (FALSE)}
}
\value{
list
}
\description{
Funkcja licząca odsetek absolwentów, którzy w danym miesiącu
(domyślnie jest to wrzesień) kontynuowali zatrudnienie u pracodawcy, który
wcześniej (tj. kiedy byli jeszcze uczniami) zatrudniał ich jako pracowników
młodocianych. Wskaźnik liczony jest tylko dla absolwentów branżowych szkół 1.
stopnia. Wyróżnianych jest 5 form kontynuowania pracy u pracodawcy
zatrudniającego poprzednio jako młodocianego:
\itemize{
\item{\code{nieucz_niekontuop}}{Nieuczący się, nie kontynuujący pracy u
danego pracodawcy}
\item{\code{nieucz_kont_uop}}{Nieuczący się, kontynuujący pracę u danego
pracodawcy na podstawie umowy o pracę}
\item{\code{nieucz_kont_inne}}{Nieuczący się, kontynuujący pracę u danego
pracodawcy w formie innej niż umowa  o pracę}
\item{\code{ucz_niekontuop}}{Uczący się, nie kontynuujący pracy u danego
pracodawcy na podstawie umowy o pracę}
\item{\code{ucz_kontuop}}{Uczący się, kontynuujący pracę u danego
pracodawcy na podstawie umowy o pracę}
}
}
