% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{E2_nauka_kontyn}
\alias{E2_nauka_kontyn}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
E2_nauka_kontyn(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
list
}
\description{
Funkcja licząca wskaźnik E2 - Sposoby kontynuowania edukacji.
Domyślnym okresem, w którym sprawdzany jest odsetek kontynuujących naukę jest
grudzień roku ukończenia szkoły (zmienna \code{p4$typ_szk_kont6}).
Od bieżącej edycji monitoringu informacja ta dołączona jest do tabeli p4 jako
zmienna \code{p4$typ_szk_kont6}, gdzie wartość "6" oznacza 6 miesięcy od
ukończenia szkoły (6 miesięcy od czerwca, czyli grudzień).
}
