% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabele_posrednie.R
\name{rozdziel_tabele}
\alias{rozdziel_tabele}
\title{Funkcje przekształcające surowe tabele pośrednie - Monitoring Karier}
\usage{
rozdziel_tabele(parametry, rok_ukonczenia = 2024)
}
\arguments{
\item{parametry}{lista zawierająca ścieżki do plików/katalogów}

\item{rok_ukonczenia}{rok, którym absolwent ukończył szkołę (jest to tym
samym rok monitoringu)}
}
\value{
obiekt \code{.RData} z tabelami pośrednimi
}
\description{
Funkcja rozłącza obiekt z tabelami pośrednimi przygotowanymi przez Tomka na
oddzielne obiekty oraz dokonuje drobnych testów przed rozpoczęciem
przygotowywania tabel do dalszych etapów pracy.
}
