% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{status_S3_mies}
\alias{status_S3_mies}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
status_S3_mies(x, rok_od, mies_od, rok_do, mies_do)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{rok_od}{rok początku okresu, dla którego ma być policzony wskaźnik}

\item{mies_od}{miesiąc początku okresu, dla którego ma być policzony wskaźnik}

\item{rok_do}{rok końca okresu, dla którego ma być policzony wskaźnik}

\item{mies_do}{miesiąc końca okresu, dla którego ma być policzony wskaźnik}
}
\value{
list
}
\description{
Funkcja licząca odsetek absolwentów o danym statusie
edukacyjno-zawodowym (wskaźnik S3) w danym miesiącu dla raportu dla danego
roku.
}
\details{
Wyróżniamy następujące statusy (bez KUZ i KKZ):
\itemize{
\item{tylko_ucz}{Tylko nauka}
\item{ucz_prac}{Nauka i praca}
\item{tylko_prac}{Tylko praca}
\item{bezrob}{Bezrobotny}
\item{brak_danych}{Brak danych o aktywności}
}
Bezrobotni kontynuujący naukę niestacjonarnie nie są wykluczani z kategorii
„Nauka”, więc określając ich status w danym miesiącu traktujemy ich jako
osoby uczące się. W sytuacji gdy w tym samym miesiącu absolwent miał różne
statusy, przypisujemy mu „bardziej aktywny” z tych statusów (np. „naukę” a
nie „bezrobocie”).
Wszystkie powyższe statusy są od tej edycji dodawane do tabel pośrednich p3,
a ta funkcja jedynie zlicza odsetek ich występowania.
}
