% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm.R
\name{Z4_ods_prac_mies}
\alias{Z4_ods_prac_mies}
\title{Wskaźniki zagregowane dla monitoringu karier - dane administracyjne}
\usage{
Z4_ods_prac_mies(x, rok_od, mies_od = 9, rok_do, mies_do = 12, nauka)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{rok_od}{rok początku okresu, dla którego ma być policzony wskaźnik}

\item{mies_od}{miesiąc początku okresu, dla którego ma być policzony wskaźnik}

\item{rok_do}{rok końca okresu, dla którego ma być policzony wskaźnik}

\item{mies_do}{miesiąc końca okresu, dla którego ma być policzony wskaźnik}

\item{nauka}{wartość TRUE/FALSE określająca czy status ma być liczony dla
absolwentów uczących się czy nie uczących się}
}
\value{
list
}
\description{
Funkcja licząca na potrzeby szablonu raportu odsetek pracujących
w danym miesiącu w podziale na pobierających i nie pobierających nauki.
Domyślnie jest to okres między wrześniem a grudniem roku przekazanego w
argumentach funkcji, ale te wartości można zmieniać.
}
